update: All Api data to String

+++ app/src/main/java/com/marinoariasg/conduentweather/screens/searchLocation/SearchByCityId.kt
@@ -8,7 +8,7 @@ class SearchByCityId(var cityId: String = "", units: String) : Search(units) {
 
     override suspend fun getDataFromRepository(repository: WeatherRepository): WeatherData {
         Timber.i("Get data searching by city Id")
-        return repository.weatherByCityId(cityId = cityId.toInt(), units = units)
+        return repository.weatherByCityId(cityId = cityId, units = units)
     }
 
 }
